fix get network id retrry behavior on windows
